discord/scheduled-discord-messages/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.9.1